+++ acceptance-tests/build.gradle
@@ -19,6 +19,7 @@ dependencies {
   testRuntimeOnly 'org.bouncycastle:bcpkix-jdk15on'
 
   testImplementation project(':ethsigner:core')
+  testImplementation project(':ethsigner:app')
   testImplementation (group: 'tech.pegasys.signers.internal', name: 'signing-secp256k1-impl', classifier: 'test-fixtures')
   testImplementation (group: 'tech.pegasys.signers.internal', name: 'keystorage-hashicorp', classifier: 'test-fixtures')
 
@@ -41,6 +42,7 @@ task acceptanceTest(type: Test) {
   description = 'Runs EthSigner acceptance tests.'
   group = 'verification'
 
+  systemProperty 'acctests.runEthSignerAsProcess', 'true'
   useJUnitPlatform()
   // toggle to show standard out and standard error of the test JVM(s) on the console
   testLogging.showStandardStreams = false
